all: use spanName as value for tag:method
For proper distinguishment from other running apps,
and later correlation:

the tag:method having a value of:
  "executeQuery"
for PreparedStatement and Statement and others
doesn't help distinguish between them. Hence
for proper debugging the fully qualified names
are distinguishable e.g.
    "java.sql.PreparedStatement.executeQuery"
    "java.sql.Statement.executeQuery"

Therefore, this change removes previous tag value
"method" that just consisted of a clashing/non-distinguishing
value and replaces it will the fully qualified name that
we also share with the span.

Fixes #42